Changed amount types to decimal. Read db data instead of hardcoded mockup.

+++ FinanceManager.API/FinanceManager.Web/Models/MonthOperationViewModel.cs
@@ -6,9 +6,9 @@ namespace FinanceManager.Web.Models
     public class MonthOperationViewModel
     {
         public string Name { get; set; }
-        public double TotalAmount { get; set; }
-        public double? AlreadyPayedAmount { get; set; }
-        public double? PaymentLeftAmount { get { return TotalAmount - AlreadyPayedAmount; } }
+        public decimal TotalAmount { get; set; }
+        public decimal? AlreadyPayedAmount { get; set; }
+        public decimal? PaymentLeftAmount { get { return TotalAmount - AlreadyPayedAmount; } }
         public DateTime FinishDate { get; set; }
